frontal_prod/morfologico/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
torch==2.1.0
torchvision==0.16.0
aiofiles==24.1.0
pydantic==2.5.0
